@gmriggs 2018-10-06
Looks good, locked to 60 fps